+++ exo-go/src/app_setup/helpers.go
@@ -1,26 +1,12 @@
 package appSetup
 
 import (
-	"io/ioutil"
 	"os"
 	"path"
 
-	yaml "gopkg.in/yaml.v2"
-
 	"github.com/Originate/exosphere/exo-go/src/types"
 	"github.com/Originate/exosphere/exo-go/test_helpers"
-	"github.com/pkg/errors"
 )
-
-func joinDockerConfigMaps(maps ...map[string]types.DockerConfig) map[string]types.DockerConfig {
-	result := map[string]types.DockerConfig{}
-	for _, currentMap := range maps {
-		for key, val := range currentMap {
-			result[key] = val
-		}
-	}
-	return result
-}
 
 // CheckoutApp copies the example app appName to cwd
 func CheckoutApp(cwd, appName string) error {
@@ -33,16 +19,12 @@ func CheckoutApp(cwd, appName string) error {
 	return testHelpers.CopyDir(src, dest)
 }
 
-// GetDockerCompose reads docker-compose.yml at the given location and
-// returns the dockerCompose object
-func GetDockerCompose(dockerComposeLocation string) (result types.DockerCompose, err error) {
-	yamlFile, err := ioutil.ReadFile(dockerComposeLocation)
-	if err != nil {
-		return result, err
+func joinDockerConfigMaps(maps ...map[string]types.DockerConfig) map[string]types.DockerConfig {
+	result := map[string]types.DockerConfig{}
+	for _, currentMap := range maps {
+		for key, val := range currentMap {
+			result[key] = val
+		}
 	}
-	err = yaml.Unmarshal(yamlFile, &result)
-	if err != nil {
-		return result, errors.Wrap(err, "Failed to unmarshal docker-compose.yml")
-	}
-	return result, nil
+	return result
 }
