+++ exo-go/src/docker_helpers/index.go
@@ -5,10 +5,14 @@ import (
 	"io/ioutil"
 	"strings"
 
+	yaml "gopkg.in/yaml.v2"
+
 	"github.com/Originate/exosphere/exo-go/src/process_helpers"
-	"github.com/docker/docker/api/types"
+	"github.com/Originate/exosphere/exo-go/src/types"
+	dockerTypes "github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/filters"
 	"github.com/docker/docker/client"
+	"github.com/pkg/errors"
 )
 
 // CatFileInDockerImage reads the file fileName inside the given image
@@ -20,12 +24,26 @@ func CatFileInDockerImage(c *client.Client, image, fileName string) ([]byte, err
 	return []byte(output), err
 }
 
+// GetDockerCompose reads docker-compose.yml at the given location and
+// returns the dockerCompose object
+func GetDockerCompose(dockerComposeLocation string) (result types.DockerCompose, err error) {
+	yamlFile, err := ioutil.ReadFile(dockerComposeLocation)
+	if err != nil {
+		return result, err
+	}
+	err = yaml.Unmarshal(yamlFile, &result)
+	if err != nil {
+		return result, errors.Wrap(err, "Failed to unmarshal docker-compose.yml")
+	}
+	return result, nil
+}
+
 // ListRunningContainers passes a slice of the names of running containers
 // and error (if any) to the callback function
 func ListRunningContainers(c *client.Client) ([]string, error) {
 	containerNames := []string{}
 	ctx := context.Background()
-	containers, err := c.ContainerList(ctx, types.ContainerListOptions{})
+	containers, err := c.ContainerList(ctx, dockerTypes.ContainerListOptions{})
 	if err != nil {
 		return containerNames, err
 	}
@@ -38,7 +56,7 @@ func ListRunningContainers(c *client.Client) ([]string, error) {
 // PullImage pulls the given image from DockerHub, returns an error if any
 func PullImage(c *client.Client, image string) error {
 	ctx := context.Background()
-	stream, err := c.ImagePull(ctx, image, types.ImagePullOptions{})
+	stream, err := c.ImagePull(ctx, image, dockerTypes.ImagePullOptions{})
 	if err != nil {
 		return err
 	}
@@ -51,7 +69,7 @@ func RemoveDanglingImages(c *client.Client) error {
 	ctx := context.Background()
 	filtersArgs := filters.NewArgs()
 	filtersArgs.Add("dangling", "true")
-	imageSummaries, err := c.ImageList(ctx, types.ImageListOptions{
+	imageSummaries, err := c.ImageList(ctx, dockerTypes.ImageListOptions{
 		All:     false,
 		Filters: filtersArgs,
 	})
@@ -59,7 +77,7 @@ func RemoveDanglingImages(c *client.Client) error {
 		return err
 	}
 	for _, imageSummary := range imageSummaries {
-		_, err = c.ImageRemove(ctx, imageSummary.ID, types.ImageRemoveOptions{})
+		_, err = c.ImageRemove(ctx, imageSummary.ID, dockerTypes.ImageRemoveOptions{})
 		if err != nil {
 			return err
 		}
